feat: #35 dialog 파일 추가
