+++ src/page/Home/Detail/index.tsx
@@ -3,83 +3,14 @@ import { useQuery } from "react-query";
 import { useNavigate, useParams } from "react-router-dom";
 import { useRecoilState, useRecoilValue } from "recoil";
 
-import { AvatarColor, HEADER_FN } from "../../../constants";
+import { HEADER_FN } from "../../../constants";
 import { recoil_ } from "../../../recoil";
 
 import { StyledDetail } from "./styled";
-import { MoreOutlined, EnvironmentOutlined } from "@ant-design/icons";
-import Avatar from "boring-avatars";
+
 import { LoadingBox2 } from "../../../components/LoadingContainer";
-import { getFormattedDate } from "../../../components/api/getFormattedDate";
 import { GetPostOne } from "../Map/api/getPost";
-import { GetRevGeocode } from "../Map/api/getRevGeocode";
-
-const DetailCard = (item: any) => {
-  const { data: geoData, isSuccess: geoSuccess } = useQuery(
-    "getRevGeo",
-    () =>
-      item &&
-      GetRevGeocode({ lat: item.data.latitude, lng: item.data.longitude }),
-    {
-      retry: 1,
-    }
-  );
-  const time = getFormattedDate(new Date(item.data.created_time));
-
-  return (
-    <>
-      {geoSuccess ? (
-        <div className="postContainer" onClick={(e) => e.stopPropagation()}>
-          <div
-            className="background"
-            style={{ background: `url(${item.data.image[0].img_url})` }}
-          ></div>
-          <div className="content">
-            <div className="header">
-              <div className="title">
-                <Avatar
-                  size={32}
-                  variant="beam"
-                  name={item.data.post_id}
-                  colors={AvatarColor}
-                />
-                username
-              </div>
-              <div className="HederAction">
-                <MoreOutlined />
-              </div>
-            </div>
-            {item.data.content}
-            <div className="PostInfo">
-              <div className="date">{time}</div>
-              <div className="location">
-                <EnvironmentOutlined /> {geoData}
-              </div>
-            </div>
-          </div>
-
-          <div className="actionSpace">
-            <div className="emotions">
-              <div className="emotionButton">
-                🥰<div className="emotionCount">9</div>
-              </div>
-              <div className="divider"></div>
-              <div className="emotionButton">
-                😎<div className="emotionCount">10</div>
-              </div>
-              <div className="divider"></div>
-              <div className="emotionButton">
-                😢<div className="emotionCount">5</div>
-              </div>
-            </div>
-          </div>
-        </div>
-      ) : (
-        <LoadingBox2 />
-      )}
-    </>
-  );
-};
+import { DetailCard } from "./components/detailCard";
 
 export const Detail = () => {
   const [dataLoading, setDataLoading] = useState(false);
@@ -95,7 +26,6 @@ export const Detail = () => {
     {
       retry: 1,
       refetchOnReconnect: false,
-      onSuccess: (res) => {},
     }
   );
